[RF] Move `ModelConfig` from RooStats to RooFit
There is a dependency problem between the RooFit packages that will
start to annoy us soon. Right now, RooFitHS3 depends on both RooFitCore,
RooFit, and RooStats.

But to really have nice interfaces for the JSON IO that can be used from
the workspace directly (like `ws.writeToFile("ws.json")`), RooFitHS3
needs to be merged with RooFitCore at some point.

As of now, this would not work, because then there would be a circular
dependency between RooStats and RooFitCore.

This can be avoided by moving the `RooStats::ModelConfig` class from
RooStats to RooFitCore, avoiding the dependency of RooFitHS3 on
RooStats.

Some arguments to make for this case:

  * The ModelConfig is strongly connected to the RooWorkspace, because
    people often use it to save model "metadata" in the workspace

  * RooFit is often said to be for "model building" and RooStats for
    statisitcal procedures. Specifying model details in the
    configuration object is part of model building.

  * Also HistFactory depends on RooStats, but only for the ModelConfig
    and the Asimov dataset generation. By moving the ModelConfig to
    RooFit, we are also half way done avoiding the dependency of
    HistFactory on RooStats.

  * The only thing that people use from RooStats is often only the
    ModelConfig anyway, because it lets them specify analysis details in
    the workspace, and then the analysis is done with other frameworks.
