+++ roofit/roofitcore/inc/LinkDef.h
@@ -345,6 +345,7 @@
 #pragma link C++ class RooMoment+ ;
 #pragma link C++ class RooFirstMoment+ ;
 #pragma link C++ class RooSecondMoment+ ;
+#pragma link C++ class RooStats::ModelConfig+;
 #pragma link C++ class RooStudyManager+ ;
 #pragma link C++ class RooStudyPackage+ ;
 #pragma link C++ class RooAbsStudy+ ;
